+++ benchmark/count_rotations.cpp
@@ -156,7 +156,7 @@ public:
 	void
 	run(int seed)
 	{
-		std::mt19937 rnd(seed);
+		std::mt19937 rnd(static_cast<unsigned long>(seed));
 		std::uniform_int_distribution distr(std::numeric_limits<int>::min(),
 		                                    std::numeric_limits<int>::max());
 
@@ -223,7 +223,7 @@ public:
 	void
 	run(int seed)
 	{
-		std::mt19937 rnd(seed);
+		std::mt19937 rnd(static_cast<unsigned long>(seed));
 		std::uniform_int_distribution distr(std::numeric_limits<int>::min(),
 		                                    std::numeric_limits<int>::max());
